[FEAT] 팝업뷰 상단을 드래그시 닫히도록 하는 BottomSheetViewController class 생성
